The TrackAssociator is now equipped to access and utilize conditions data in a generic way. This is necessary for the TrackAssociator to determine if a given DT or CSC chamber should or should not be used for matching because it is marked as bad, and if bad or noisy calorimeter cells should be used in energy deposits, etc. The change is particularly vital at the time of this commit because the ME4/2 chambers were added to the geometry, and the Track- Associator needs to be able to figure out that nearly all are marked as bad in the conditions data, and hence should not be used for matching.
Several files were modified to achieve the above. Currently the only specif-
ic conditions-data-checking implementation is for the CSCs to address the
immediate problems just mentioned. We access the CSCBadChambers object from
the CSCBadChambersRcd, and check each chamber in the geometry for presence
in CSCBadChambers before including the chamber in the list of valid detector
elements to use in matching. By default, bad chambers are not included in
the matching.